bin/requirements.txt
annotated-types==0.7.0
apache-beam==2.60.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cloudpickle==2.2.1
comm==0.2.2
contourpy==1.3.0
crcmod==1.7
cryptography==43.0.3
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.1.1
dnspython==2.7.0
docopt==0.6.2
docstring_parser==0.16
exceptiongroup==1.2.2
executing==2.1.0
facets-overview==1.1.1
fastavro==1.9.7
fasteners==0.19
fastjsonschema==2.20.0
fonttools==4.54.1
future==1.0.0
google-api-core==2.21.0
google-apitools==0.5.31
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.70.0
google-cloud-bigquery==3.26.0
google-cloud-bigquery-storage==2.27.0
google-cloud-bigtable==2.26.0
google-cloud-core==2.4.1
google-cloud-dataproc==5.13.0
google-cloud-datastore==2.20.1
google-cloud-dlp==3.23.0
google-cloud-language==2.14.0
google-cloud-pubsub==2.26.1
google-cloud-pubsublite==1.11.1
google-cloud-recommendations-ai==0.10.12
google-cloud-resource-manager==1.12.5
google-cloud-spanner==3.49.1
google-cloud-storage==2.18.2
google-cloud-videointelligence==2.13.5
google-cloud-vision==3.7.4
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpc-interceptor==0.15.4
grpcio==1.65.5
grpcio-status==1.62.3
hdfs==2.7.3
httplib2==0.22.0
idna==3.10
importlib_metadata==8.4.0
ipykernel==6.29.5
ipython==8.28.0
ipywidgets==8.1.5
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.4
jsonpickle==3.3.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.2.0
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.13
keyring==25.4.1
keyrings.google-artifactregistry-auth==1.1.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mistune==3.0.2
more-itertools==10.5.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==1.26.4
oauth2client==4.1.3
objsize==0.7.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
orjson==3.10.9
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==4.25.5
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydot==1.4.2
Pygments==2.18.0
pymongo==4.10.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
redis==5.1.1
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rpds-py==0.20.0
rsa==4.9
SecretStorage==3.3.3
shapely==2.0.6
six==1.16.0
soupsieve==2.6
sql==2022.4.0
sqlparse==0.5.1
stack-data==0.6.3
timeloop==1.0.2
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
webencodings==0.5.1
widgetsnbextension==4.0.13
wrapt==1.16.0
zipp==3.20.2
zstandard==0.23.0